+++ nodescraper/plugins/inband/niccli/nic_data.py
